// frontend/web/src/services/API/proserve-wb-packaging-api/models/UpdatePipelineRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdatePipelineRequest
 */
export interface UpdatePipelineRequest {
    /**
     * The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build
     * @type {Array<string>}
     * @memberof UpdatePipelineRequest
     */
    buildInstanceTypes?: Array<string>;
    /**
     * Pipeline ID
     * @type {string}
     * @memberof UpdatePipelineRequest
     */
    pipelineId: string;
    /**
     * The schedule of the pipeline. See https://docs.aws.amazon.com/imagebuilder/latest/userguide/cron-expressions.html
     * @type {string}
     * @memberof UpdatePipelineRequest
     */
    pipelineSchedule?: string;
    /**
     * Recipe version ID of the recipe version associated to the pipeline
     * @type {string}
     * @memberof UpdatePipelineRequest
     */
    recipeVersionId?: string;
    /**
     * Product ID to associate with the pipeline for automatic version creation
     * @type {string}
     * @memberof UpdatePipelineRequest
     */
    productId?: string;
}

/**
 * Check if a given object implements the UpdatePipelineRequest interface.
 */
export function instanceOfUpdatePipelineRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "pipelineId" in value;

    return isInstance;
}

export function UpdatePipelineRequestFromJSON(json: any): UpdatePipelineRequest {
    return UpdatePipelineRequestFromJSONTyped(json, false);
}

export function UpdatePipelineRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdatePipelineRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'buildInstanceTypes': !exists(json, 'buildInstanceTypes') ? undefined : json['buildInstanceTypes'],
        'pipelineId': json['pipelineId'],
        'pipelineSchedule': !exists(json, 'pipelineSchedule') ? undefined : json['pipelineSchedule'],
        'recipeVersionId': !exists(json, 'recipeVersionId') ? undefined : json['recipeVersionId'],
        'productId': !exists(json, 'productId') ? undefined : json['productId'],
    };
}

export function UpdatePipelineRequestToJSON(value?: UpdatePipelineRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'buildInstanceTypes': value.buildInstanceTypes,
        'pipelineId': value.pipelineId,
        'pipelineSchedule': value.pipelineSchedule,
        'recipeVersionId': value.recipeVersionId,
        'productId': value.productId,
    };
}