// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetPipelineResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Pipeline } from './Pipeline';
import {
    PipelineFromJSON,
    PipelineFromJSONTyped,
    PipelineToJSON,
} from './Pipeline';

/**
 * 
 * @export
 * @interface GetPipelineResponse
 */
export interface GetPipelineResponse {
    /**
     * 
     * @type {Pipeline}
     * @memberof GetPipelineResponse
     */
    pipeline: Pipeline;
}

/**
 * Check if a given object implements the GetPipelineResponse interface.
 */
export function instanceOfGetPipelineResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "pipeline" in value;

    return isInstance;
}

export function GetPipelineResponseFromJSON(json: any): GetPipelineResponse {
    return GetPipelineResponseFromJSONTyped(json, false);
}

export function GetPipelineResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetPipelineResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pipeline': PipelineFromJSON(json['pipeline']),
    };
}

export function GetPipelineResponseToJSON(value?: GetPipelineResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pipeline': PipelineToJSON(value.pipeline),
    };
}